fix(.emacs): handle missing `restart-emacs` and `RESTART` arg
`restart-emacs` was introduced in `29.x`. That isn't necessarily a problem,
except it's just a wrapper around `save-buffers-kill-emacs`, which in the end
calls `kill-emacs` C-func. _That_ func has a new `RESTART` argument starting
with `29.x` and that's a problem.

For the time being, let it pass to the `(error)` and crash.