refactor(.emacs,elpaca): make installer-restarter work on <29.1
Technically, it doesn't restart. It unloads some variables `package.el` filled
during the init. Normally, the `(restart-emacs)` in this case is supposed to
unload features and ensure these are loaded from Elpaca folders.

That's tested later on in the installation behavior tests, so slightly
unnecessary to restart, but keeps the edge-case of a post-init running Emacs
hanging around. Will be solved later, if needed.