fix(.emacs): wrap problematic GUI-based functionality
Some functions from `window.el` and friends get loaded only when a real window
is provided. That, unfortunately causes shenanigans when trying to have these
funcs around and wanting to compile the file.

The simple wrapper only makes the compiler quiet because it'll call the
problematic func internally or crash. Almost the same thing, but no annoying
false-positive errors.