fix(bindings): disable `multi-term-leftside` test on `27.1` and `27.2`
Due to some calculation or missing attribute set on non-interactive the
behavior is untestable in a container due to `(split-window-right)` call
causing an error of "Window %s too small for splitting". This does NOT happen
for `(split-window-below)`.

An alternative of `(split-window nil 2 'right)` works but then a re-calculation
through `(balance-windows)` doesn't due to (possibly) the same missing
attribute set throwing `(wrong-type-argument number-or-marker-p nil)`
confirming it's about "something" not being set when the window isn't
real/rendered.